+++ Lib/importlib/resources/_adapters.py
