+++ Lib/importlib/resources/_common.py
