+++ Lib/importlib/resources/_itertools.py
