+++ Lib/importlib/resources/_legacy.py
