+++ Misc/NEWS.d/next/Library/2021-12-17-16-27-44.bpo-46118.euAy0E.rst
@@ -1,0 +1,1 @@
+Moved importlib.resources and its related functionality to a package.
